requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.5
redis[hiredis]==5.2.1
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.4
